feat: replace mocha-phantomjs with mocha-headless-chrome (#57)
Updated the tooling to use mocha-headless-chrome instead of mocha-phantomjs. In addition added a new target for running the tests specifically on the CI, as the headless environment requires using the chromium browser installed on the host os as the binary pulled via npm doesn't work. This change is intended to be temporary, until the tests can be broken into seperate unit/ux tests.
Co-authored-by: Luke <lsh-0@users.noreply.github.com>